Changes to decrease reliance on timing in unit tests: * Wait to connect to ZK in gobblin.yarn.GobblinYarnAppLauncherTest.setUp() * Add AssertWithBackoff to retry an assert multiple times * Refactor tests gobblin-yarn/src/test/java/gobblin/yarn/GobblinApplicationMasterTest.java and gobblin-yarn/src/test/java/gobblin/yarn/GobblinYarnAppLauncherTest.java to use AssertWithBackoff

+++ gobblin-scheduler/src/test/java/gobblin/scheduler/JobConfigFileMonitorTest.java
@@ -20,16 +20,20 @@ import java.util.Set;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.testng.Assert;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
+import com.google.common.base.Function;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
 import com.google.common.util.concurrent.ServiceManager;
 
 import gobblin.configuration.ConfigurationKeys;
+import gobblin.testing.AssertWithBackoff;
 
 
 /**
@@ -45,6 +49,15 @@ public class JobConfigFileMonitorTest {
   private ServiceManager serviceManager;
   private JobScheduler jobScheduler;
   private File newJobConfigFile;
+
+  private class GetNumScheduledJobs implements Function<Void, Integer> {
+
+    @Override
+    public Integer apply(Void input) {
+      return JobConfigFileMonitorTest.this.jobScheduler.getScheduledJobs().size();
+    }
+
+  }
 
   @BeforeClass
   public void setUp()
@@ -63,9 +76,9 @@ public class JobConfigFileMonitorTest {
   @Test
   public void testAddNewJobConfigFile()
       throws Exception {
-    Thread.sleep(2000);
-
-    Assert.assertEquals(this.jobScheduler.getScheduledJobs().size(), 3);
+    final Logger log = LoggerFactory.getLogger("testAddNewJobConfigFile");
+    AssertWithBackoff assertWithBackoff = AssertWithBackoff.create().logger(log).timeoutMs(5000);
+    assertWithBackoff.assertEquals(new GetNumScheduledJobs(), 3, "3 scheduled jobs");
 
     // Create a new job configuration file by making a copy of an existing
     // one and giving a different job name
@@ -75,7 +88,7 @@ public class JobConfigFileMonitorTest {
     this.newJobConfigFile = new File(JOB_CONFIG_FILE_DIR, "Gobblin-test-new.pull");
     jobProps.store(new FileWriter(this.newJobConfigFile), null);
 
-    Thread.sleep(2000);
+    assertWithBackoff.assertEquals(new GetNumScheduledJobs(), 4, "4 scheduled jobs");
 
     Set<String> jobNames = Sets.newHashSet(this.jobScheduler.getScheduledJobs());
     Assert.assertEquals(jobNames.size(), 4);
@@ -89,6 +102,7 @@ public class JobConfigFileMonitorTest {
   @Test(dependsOnMethods = {"testAddNewJobConfigFile"})
   public void testChangeJobConfigFile()
       throws Exception {
+    final Logger log = LoggerFactory.getLogger("testChangeJobConfigFile");
     Assert.assertEquals(this.jobScheduler.getScheduledJobs().size(), 4);
 
     // Make a change to the new job configuration file
@@ -97,7 +111,8 @@ public class JobConfigFileMonitorTest {
     jobProps.setProperty(ConfigurationKeys.JOB_COMMIT_POLICY_KEY, "partial");
     jobProps.store(new FileWriter(this.newJobConfigFile), null);
 
-    Thread.sleep(2000);
+    AssertWithBackoff.create().logger(log).timeoutMs(5000)
+        .assertEquals(new GetNumScheduledJobs(), 4, "4 scheduled jobs");
 
     Set<String> jobNames = Sets.newHashSet(this.jobScheduler.getScheduledJobs());
     Assert.assertEquals(jobNames.size(), 4);
@@ -111,6 +126,7 @@ public class JobConfigFileMonitorTest {
   @Test(dependsOnMethods = {"testChangeJobConfigFile"})
   public void testUnscheduleJob()
       throws Exception {
+    final Logger log = LoggerFactory.getLogger("testUnscheduleJob");
     Assert.assertEquals(this.jobScheduler.getScheduledJobs().size(), 4);
 
     // Disable the new job by setting job.disabled=true
@@ -119,7 +135,9 @@ public class JobConfigFileMonitorTest {
     jobProps.setProperty(ConfigurationKeys.JOB_DISABLED_KEY, "true");
     jobProps.store(new FileWriter(this.newJobConfigFile), null);
 
-    Thread.sleep(2000);
+
+    AssertWithBackoff.create().logger(log).timeoutMs(5000)
+        .assertEquals(new GetNumScheduledJobs(), 3, "3 scheduled jobs");
 
     Set<String> jobNames = Sets.newHashSet(this.jobScheduler.getScheduledJobs());
     Assert.assertEquals(jobNames.size(), 3);
@@ -131,7 +149,7 @@ public class JobConfigFileMonitorTest {
   @AfterClass
   public void tearDown()
       throws TimeoutException {
-    this.newJobConfigFile.delete();
+    if (null != this.newJobConfigFile) this.newJobConfigFile.delete();
     this.serviceManager.stopAsync().awaitStopped(5, TimeUnit.SECONDS);
   }
 }
